### 2016-04-11 #34 @Dahaniel
The Database of Odorant Responses (DoOR.functions)
- 1. What does this package do? (explain in 50 words or less)

Package containing the functions for the DoOR project. DoOR provides a framework to merge heterogeneous response data into a single consensus database, thus creating a comprehensive view on olfactory coding.
- 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: DoOR.functions
Title: Integrating Heterogeneous Odorant Response Data into a Common Response
    Model: A DoOR to the Complete Olfactome
Version: 2.0.0
    person("Daniel", "Münch", email = "daniel.muench@uni-konstanz.de", role = c("aut", "cre")),
    person("C. Giovanni", "Galizia", email = "giovanni.galizia@uni-konstanz.de", role = "aut"),
    person("Shouwen", "Ma", role = "aut"),
    person("Martin", "Strauch", role = "aut"),
    person("Anja", "Nissler", role = "aut"),
    person("Wolf", "Huetteroth", role = "ctb"))
Description: This is a function package providing functions to perform data
    manipulations and visualizations for DoOR.data. See http://dx.doi.org/10.1093/chemse/bjq042 for the original publication and http://dx.doi.org/10.1038/srep21841
    regarding version 2.0.
    utils
    DoOR.data (>= 2.0.0),
Remotes: Dahaniel/DoOR.data@v2.0.0
    grid,
    gridExtra (>= 2.0.0),
    class,

```
- 3. URL for the package (the development repository, not a stylized html page)

https://github.com/Dahaniel/DoOR.functions
- 4. What data source(s) does it work with (if applicable)?

It provides its own data via the `DoOR.data` package (see corresponding onboarding issue).
- 5. Who is the target audience?

Scientists working on olfactory coding, modelers and physiologists.
- 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?

No, to our knowledge DoOR is the only project providing this kind of data and using this approach to integrate heterogeneous data.
- 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
  - [x] This package does not violate the Terms of Service of any service it interacts with.
  - [x] The repository has continuous integration with Travis CI and/or another service
  - [x] The package contains a vignette
  - [x] The package contains a reasonably complete README with `devtools` install instructions
  - [ ] The package contains unit tests
  - [x] The package only exports functions to the NAMESPACE that are intended for end users
- 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
  - [ ] Are there any package dependencies not on CRAN? - _no_
  - [x] Does the package have a CRAN accepted license?
  - [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below. - _no_
- 9. Please add explanations below for any exceptions to the above:
  - Not sure how to implement unit tests but am willing to learn, might need some help.
- 10. If this is a resubmission following rejection, please explain the change in circumstances.
