### 2016-05-25 #34 @RemkoDuursma
Apologies for my very slow review.

- The DoOR.functions package is quite well documented and implemented with good consistency. I did not find any major issues here at all. Again I want to point out that I cannot comment on the usefulness or applicability of the package since this is well outside my expertise.
- The included vignettes are very useful to get an overview of this rather complicated package.
- The `default.val` function is slightly odd, and could be replaced with the more standard use of `options` to set values for default parameters/flags in a package. The drawback of the current implementation is that new default values cannot be set, this would be a useful feature I think. One option is to have all possible parameters as arguments to the `default.val` function (e.g. `points.cex=1`).
- In the DESCRIPTION file, it is customary to move all URLs to the URL field (there is one in the Description field).
- Names of functions use camelCase which is fine (though snake_case is preferred by ROpenSci as previously discussed), except `CreateDatabase`.
- The package includes examples, but it would be helpful if the examples are commented (at the moment only the code is shown but no information on what the code does).
- Some functions lack in documentation (e.g. `updateOdorInfo`).
