### 2016-08-10 #34 @sckott
@Dahaniel can you respond to the review above please
