+++ DESCRIPTION
@@ -10,11 +10,10 @@ Authors@R: c(
     person("Shouwen", "Ma", role = "aut"),
     person("Martin", "Strauch", role = "aut"),
     person("Anja", "Nissler", role = "aut"))
-URL: http://neuro.uni-konstanz.de/DoOR
+URL: http://neuro.uni.kn/DoOR, http://dx.doi.org/10.1038/srep21841, http://dx.doi.org/10.1093/chemse/bjq042
 BugReports: https://github.com/Dahaniel/DoOR.data/issues
 Description: This is a data package providing Drosophila odorant response data for
-    DoOR.functions. See http://dx.doi.org/10.1093/chemse/bjq042 for the original
-    publication and http://dx.doi.org/10.1038/srep21841 regarding version 2.0.
+    DoOR.functions. See URLs for the original and the DoOR 2.0 publications.
 License: CC BY-SA 4.0
 Encoding: UTF-8
 Imports:
