### 2016-08-11 1f5b93a
clarified where loadData() loads data to
