### 2016-08-11 #34 @Dahaniel
@RemkoDuursma Thank you very much for reviewing the package and apologies for my slow response.
- It is correct that the `default.val()` values cannot be changed. `default.val()` is intended to provide package wide default values for functions, whenever these are called without parameters. In fact users can change the default parameters when calling the individual functions with parameters. So I think changing the values that `default.val()` provides would not make sense.
Regarding the suggestion to use `options` instead: I was not aware of `options` and have to read about its usage. When this would provide a standard way of achieving what we achieve with `default.val()` we should use `options` instead.

We also use `default.val` to check if the data.frames a function relies on exist in the global environment and if not load the data from the package:
```r
  if (DoOR_default == "DoOR_mappings") {
    if (!exists("DoOR.mappings")) {
      data(DoOR.mappings)
    }
    return(DoOR.mappings)
  }
```
The idea is that the user can put modified data into the global environment that will be used when available. Not sure whether this could be implemented with options.
- I moved the urls to the URL field
- Function names: see my comment at the DoOR.data thread. I am happy to change function names having a transitional phase where old and new name work and am looking for a way to implement this.
- Documentation: I will check for missing docs and add descriptions to examples
