### 2016-08-11 #34 @sckott
thanks for your response @Dahaniel  on function names, see comment in #35
