### 2016-09-16 cdaa244
Added warning that loadData() loads into current environment.
