+++ man/load_door_data.Rd
@@ -1,10 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/loadData.R
-\name{loadData}
+\name{load_door_data}
 \alias{loadData}
-\title{loadData}
+\alias{load_door_data}
+\title{load_door_data}
 \usage{
-loadData()
+load_door_data()
 }
 \value{
 attaches all DoOR data to the main workspace
@@ -14,7 +15,7 @@ load all DoOR.data
 }
 \examples{
 
-loadData()
+load_door_data()
 
 }
 \author{
