+++ man/AL.map.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DoOR.data-package.R
 \docType{data}
-\name{AL.map}
-\alias{AL.map}
-\title{AL.map}
+\name{door_AL_map}
+\alias{door_AL_map}
+\title{door_AL_map}
 \format{The format is: List of 5 $ glomeruli :'data.frame': 19834 obs. of 3
 variables: ..$ glomerulus: Factor w/ 51 levels "D","DA1","DA2",..: 4 4 4 4 4
 4 4 4 4 4 ...  ..$ x : num [1:19834] 33.1 33.1 33.1 33.1 33.1 ...  ..$ y :
