+++ man/data.format.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DoOR.data-package.R
 \docType{data}
-\name{data.format}
-\alias{data.format}
-\title{data.format}
+\name{door_data_format}
+\alias{door_data_format}
+\title{door_data_format}
 \format{A data frame with 693 observations on 5 variables.}
 \description{
 DoOR response data template
