### 2017-02-24 61df467
renamed DoOR.mappings -> door_mappings
