+++ man/DoOR.mappings.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DoOR.data-package.R
 \docType{data}
-\name{DoOR.mappings}
-\alias{DoOR.mappings}
-\title{DoOR.mappings}
+\name{door_mappings}
+\alias{door_mappings}
+\title{door_mappings}
 \format{A data frame with 96 observations on 20 variables.}
 \description{
 the mappings of DoOR responding units to the correesponding morphological structures
