### 2017-02-24 bd5bc29
renamed excluded.data -> door_excluded_data
