### 2017-02-24 19cff42
renamed glo.dist -> door_glo_dist
