+++ man/glo.dist.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DoOR.data-package.R
 \docType{data}
-\name{glo.dist}
-\alias{glo.dist}
-\title{glo.dist}
+\name{door_glo_dist}
+\alias{door_glo_dist}
+\title{door_glo_dist}
 \format{A data frame with 49 observations on 49 variables.}
 \description{
 glomerular distances
