### 2017-02-24 4c363ac
renamed response.range -> door_response_range
