+++ man/response.range.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DoOR.data-package.R
 \docType{data}
-\name{response.range}
-\alias{response.range}
-\title{response.range}
+\name{door_response_range}
+\alias{door_response_range}
+\title{door_response_range}
 \format{A data frame with 42 observations on 4 variables.}
 \description{
 the response range of the studies included in DoOR
