### 2017-02-24 cd89839
renamed weight.globNorm -> door_global_normalization_weights
