### 2017-02-24 322af7d
renamed response.matrix_non.normalized -> door_response_matrix_non_normalized
