+++ man/response.matrix_non.normalized.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DoOR.data-package.R
 \docType{data}
-\name{response.matrix_non.normalized}
-\alias{response.matrix_non.normalized}
+\name{door_response_matrix_non_normalized}
+\alias{door_response_matrix_non_normalized}
 \title{non normalized DoOR response matrix}
 \format{A data frame with 693 observations on 78 variables.}
 \description{
