### 2017-02-24 8b2d31e
renamed response.matrix -> door_response_matrix
