+++ man/response.matrix.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/DoOR.data-package.R
 \docType{data}
-\name{response.matrix}
-\alias{response.matrix}
+\name{door_response_matrix}
+\alias{door_response_matrix}
 \title{DoOR response matrix}
 \format{A data frame with 693 observations on 78 variables.}
 \description{
