### 2017-03-06 bcbeb85
added data set documentation
