+++ R/DoOR.data-package.R
@@ -1,1380 +1,2347 @@
-
-
 #' ab2B
-#' 
-#' DoOR response data
-#' 
+#'
+#' DoOR response data for responding unit ab2B. Please find detailed
+#' information on the respective sources of the data in door_dataset_info.
+#'
 #' @name ab2B
 #' @docType data
-#' @format A data frame with 693 observations on 12 variables.
-#' @keywords datasets
-
-NULL
-
-
-
-
-
-#' ab4B
-#' 
-#' DoOR response data
-#' 
-#' @name ab4B
-#' @docType data
-#' @format A data frame with 693 observations on 10 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ab5B
-#' 
-#' DoOR response data
-#' 
-#' @name ab5B
-#' @docType data
-#' @format A data frame with 693 observations on 9 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac1A
-#' 
-#' DoOR response data
-#' 
-#' @name ac1A
-#' @docType data
-#' @format A data frame with 693 observations on 7 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac1BC
-#' 
-#' DoOR response data
-#' 
-#' pooled measurements of ac1B and ac1C OSNs
-#' 
-#' @name ac1BC
-#' @docType data
-#' @format A data frame with 693 observations on 6 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac1B
-#' 
-#' DoOR response data
-#' 
-#' @name ac1B
-#' @docType data
-#' @format A data frame with 693 observations on 6 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac1
-#' 
-#' DoOR response data
-#' 
-#' pooled measurements of the ac1 sensillum
-#' 
-#' @name ac1
-#' @docType data
-#' @format A data frame with 693 observations on 6 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac2A
-#' 
-#' DoOR response data
-#' 
-#' @name ac2A
-#' @docType data
-#' @format A data frame with 693 observations on 7 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac2BC
-#' 
-#' DoOR response data
-#' 
-#' pooled measurements of ac2B and ac2C OSNs
-#' 
-#' @name ac2BC
-#' @docType data
-#' @format A data frame with 693 observations on 6 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac2B
-#' 
-#' DoOR response data
-#' 
-#' @name ac2B
-#' @docType data
-#' @format A data frame with 693 observations on 6 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac2
-#' 
-#' DoOR response data
-#' 
-#' pooled measurements of the ac2 sensilum
-#' 
-#' @name ac2
-#' @docType data
-#' @format A data frame with 693 observations on 6 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac3A
-#' 
-#' DoOR response data
-#' 
-#' @name ac3A
-#' @docType data
-#' @format A data frame with 693 observations on 8 variables.
-#' @keywords datasets
-#' 
-NULL
-
-
-
-
-
-#' ac3B
-#' 
-#' DoOR response data
-#' 
-#' @name ac3B
-#' @docType data
[…]
