+++ man/door_response_matrix.Rd
@@ -1,0 +1,94 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/DoOR.data-package.R
+\docType{data}
+\name{door_response_matrix}
+\alias{door_response_matrix}
+\title{door_response_matrix}
+\format{'data.frame':	693 obs. of  78 variables:
+$ ac1A       : num  0.0627 0.0226 0.4768 0.061 NA ...
+$ ac1B       : num  0.0697 NA NA 0.1133 NA ...
+$ ac2A       : num  0.056 0.0868 0.0496 0.8839 NA ...
+$ ac2B       : num  0.0709 NA NA 0.0443 NA ...
+$ ac3A       : num  0.0686 0.2225 0.0719 0.1899 0.1655 ...
+$ ac3B       : num  0.0447 0.0612 0.3783 0.4122 0.0628 ...
+$ ac4        : num  0.0927 0.1048 0.151 0.0921 0.0867 ...
+$ Or1a       : num  0 0.0249 NA NA NA ...
+$ Or2a       : num  0.0485 NA 0.0566 0.0638 0.0519 ...
+$ Or7a       : num  0.0253 0.0944 0.0675 0.0504 0.0189 ...
+$ Or9a       : num  0.0643 0.0688 0.1084 0.0979 0.1399 ...
+$ Or10a      : num  0.0517 0.0567 0.0539 0.0567 0.0464 ...
+$ Or13a      : num  0.0798 0.1977 0.1234 0.1535 NA ...
+$ Or19a      : num  0.1001 NA 0.0972 0.0807 0.086 ...
+$ Or22a      : num  0.0845 0.368 0.185 0.1751 0.0899 ...
+$ Or22b      : num  NA NA NA NA NA NA NA NA NA NA ...
+$ Or22c      : logi  NA NA NA NA NA NA ...
+$ Or23a      : num  0.01932 NA 0.01502 0.01502 0.00429 ...
+$ Or24a      : logi  NA NA NA NA NA NA ...
+$ Or30a      : num  0.0124 0.0435 NA NA NA ...
+$ Or33a      : num  0 0.0171 NA NA NA ...
+$ Or33b      : num  0.0628 NA 0.00813 0.02421 0.0142 ...
+$ Or33c      : num  0.0456 NA NA NA NA ...
+$ Or35a      : num  0.0816 NA 0.0732 0.0599 0.0997 ...
+$ Or42a      : num  0.082 0.123 0.031 0.175 NA ...
+$ Or42b      : num  0.0516 0.2246 0.1126 0.2586 NA ...
+$ Or43a      : num  0.0953 NA 0.0907 0.0322 0.0483 ...
+$ Or43b      : num  0.0149 0.0654 0.0196 0.0493 0.0532 ...
+$ Or45a      : num  0.00866 0.25516 NA NA NA ...
+$ Or45b      : num  0.0248 0.028 NA NA NA ...
+$ Or46a      : num  0.0981 NA NA NA NA ...
+$ Or47a      : num  0.0519 NA 0.0814 0.0221 0.0237 ...
+$ Or47b      : num  0.198 NA 0.178 0.133 0.158 ...
+$ Or49a      : num  0 0.012 NA NA NA ...
+$ Or49b      : num  0.048 0.0791 0.0561 0.0641 0.0552 ...
+$ Or59a      : num  0.00684 NA NA NA NA ...
+$ Or59b      : num  0.0237 0.0427 0.0115 0.0297 0.0276 ...
+$ Or59c      : num  0.0358 0.0369 0.0348 0.0644 NA ...
+$ Or65a      : num  0.0595 NA 0.0693 0.0527 0.0561 ...
[…]
