+++ man/ac4.Rd
@@ -4,12 +4,21 @@
 \name{ac4}
 \alias{ac4}
 \title{ac4}
-\format{A data frame with 693 observations on 8 variables.}
+\format{'data.frame':	693 obs. of  8 variables:
+$ Class            : Factor w/ 17 levels "acid","acids",..: NA 13 5 5 5 5 5 5 5 5 ...
+$ Name             : Factor w/ 690 levels "11-cis vaccenyl acetate",..: 634 675 240 613 283 239 363 436 458 341 ...
+$ InChIKey         : Factor w/ 693 levels "ACCRBMDJCPPJDX-UHFFFAOYSA-N",..: 482 612 549 252 548 418 196 577 41 462 ...
+$ CID              : Factor w/ 687 levels "1001","10050",..: 686 680 139 15 220 189 483 610 564 468 ...
+$ CAS              : Factor w/ 677 levels "1001-45-2","10032-13-0",..: 676 591 212 100 379 586 231 114 62 200 ...
+$ Yao.2005.WT      : int  42 NA NA 38 NA 55 NA 35 15 49 ...
+$ Silbering.2011.WT: num  0 6.06 33.47 9 -2 ...
+$ Marshall.2010.WT : num  0 1.67 7.5 -1 NA ...}
 \description{
-DoOR response data
+DoOR response data for responding unit ac4. Please find detailed
+information on the respective sources of the data in door_dataset_info.
 }
-\details{
-pooled measurements of the ac4 sensillum
-}
-\keyword{datasets}
+\keyword{DoOR}
+\keyword{ac4}
+\keyword{dataset}
+\keyword{responding_unit}
 
