+++ man/Or83c.Rd
@@ -4,9 +4,19 @@
 \name{Or83c}
 \alias{Or83c}
 \title{Or83c}
-\format{A data frame with 693 observations on 6 variables.}
+\format{'data.frame':	693 obs. of  6 variables:
+$ Class           : Factor w/ 17 levels "acid","acids",..: NA 13 5 5 5 5 5 5 5 5 ...
+$ Name            : Factor w/ 690 levels "11-cis vaccenyl acetate",..: 634 675 240 613 283 239 363 436 458 341 ...
+$ InChIKey        : Factor w/ 693 levels "ACCRBMDJCPPJDX-UHFFFAOYSA-N",..: 482 612 549 252 548 418 196 577 41 462 ...
+$ CID             : Factor w/ 687 levels "1001","10050",..: 686 680 139 15 220 189 483 610 564 468 ...
+$ CAS             : Factor w/ 677 levels "1001-45-2","10032-13-0",..: 676 591 212 100 379 586 231 114 62 200 ...
+$ Ronderos.2014.WT: int  0 NA NA NA NA 20 -17 NA NA NA ...}
 \description{
-DoOR response data
+DoOR response data for responding unit Or83c. Please find detailed
+information on the respective sources of the data in door_dataset_info.
 }
-\keyword{datasets}
+\keyword{DoOR}
+\keyword{Or83c}
+\keyword{dataset}
+\keyword{responding_unit}
 
