+++ man/glo.dist.Rd
@@ -1,12 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/DoOR.data-package.R
-\docType{data}
-\name{door_glo_dist}
-\alias{door_glo_dist}
-\title{door_glo_dist}
-\format{A data frame with 49 observations on 49 variables.}
-\description{
-glomerular distances
-}
-\keyword{datasets}
-
