+++ man/Or85b.Rd
@@ -4,9 +4,26 @@
 \name{Or85b}
 \alias{Or85b}
 \title{Or85b}
-\format{A data frame with 693 observations on 13 variables.}
+\format{'data.frame':	693 obs. of  13 variables:
+$ Class           : Factor w/ 17 levels "acid","acids",..: NA 13 5 5 5 5 5 5 5 5 ...
+$ Name            : Factor w/ 690 levels "11-cis vaccenyl acetate",..: 634 675 240 613 283 239 363 436 458 341 ...
+$ InChIKey        : Factor w/ 693 levels "ACCRBMDJCPPJDX-UHFFFAOYSA-N",..: 482 612 549 252 548 418 196 577 41 462 ...
+$ CID             : Factor w/ 687 levels "1001","10050",..: 686 680 139 15 220 189 483 610 564 468 ...
+$ CAS             : Factor w/ 677 levels "1001-45-2","10032-13-0",..: 676 591 212 100 379 586 231 114 62 200 ...
+$ Hallem.2006.EN  : int  13 NA 28 37 42 NA NA NA NA NA ...
+$ Dobritsa.2003.EN: int  NA NA NA NA NA NA NA NA NA NA ...
+$ Marshall.2010.WT: num  0 25.1 NA 13.5 NA ...
+$ Stensmyr.2003.WT: int  NA NA NA NA NA NA NA NA NA NA ...
+$ Schmuker.2007.TR: int  8 NA NA 27 NA 25 17 NA NA NA ...
+$ Bruyne.2001.WT  : num  8 NA NA NA NA NA NA NA NA NA ...
+$ Hallem.2004.EN  : num  NA NA NA 79.5 NA NA NA NA NA NA ...
+$ Hallem.2004.WT  : num  NA NA NA 44.2 NA ...}
 \description{
-DoOR response data
+DoOR response data for responding unit Or85b. Please find detailed
+information on the respective sources of the data in door_dataset_info.
 }
-\keyword{datasets}
+\keyword{DoOR}
+\keyword{Or85b}
+\keyword{dataset}
+\keyword{responding_unit}
 
