+++ DESCRIPTION
@@ -2,8 +2,8 @@ Package: DoOR.data
 Type: Package
 Title: Integrating Heterogeneous Odorant Response Data into a Common Response
     Model: A DoOR to the Complete Olfactome
-Version: 2.0.0
-Date: 2016-02-25
+Version: 2.0.1
+Date: 2017-03-06
 Authors@R: c(
     person("Daniel", "Münch", email = "daniel.muench@uni-konstanz.de", role = c("aut", "cre")),
     person("C. Giovanni", "Galizia", email = "giovanni.galizia@uni-konstanz.de", role = "aut"), 
@@ -21,7 +21,7 @@ Imports:
 Depends:
     R (>= 3.2.0)
 Suggests:
-    DoOR.functions (>= 2.0.0)
+    DoOR.functions (>= 2.0.1)
 Remotes: 
-    Dahaniel/DoOR.functions@v2.0.0
+    Dahaniel/DoOR.functions@v2.0.1
 RoxygenNote: 5.0.1
