### 2017-03-07 589804e
Added possibility to suppress user interaction when loading data as otherwise e.g. the vignettes would not build.
