### 2017-03-07 #34 @Dahaniel
Hi everyone, apologies that it took me so long to implement the requested changes. But finally today I was able to release v2.0.1 for both packages, implementing all of the above discussed changes.
In detail (I summarize the changes here as most affect both packages):

- [x] I renamed `loadData` to `load_door_data()` to be more descriptive.
- [x] I warn that data will be loaded to the current environment and forced a user interaction for that.
- [x] I transform all function names to snake_case and inform about that on package loading. I also kept the old function names working for now.
- [x] I replaced '.' with '_' in data names.
- [x] moved all urls to url field in DESCRIPTION.
- [x] I added documentation to the data sets.
- [x] I added some descriptions to the examples.
- [x] I exported all data as csv for better change tracking.
- [x] I kept `default.val()` for now (see previous discussion)
