### 2017-03-07 #34 @sckott
thx @Dahaniel

@RemkoDuursma if possible, could you take a look at the changes and let us know if you're happy with the changes?  also let us know if you're too busy/etc, and we'll have a look instead
