### 2017-04-19 #34 @Dahaniel
Thanks for the comments @sckott

> * Looks like you still don't have unit tests - please add those

That is right, as explained before I do not have a clue about unit testing and
might need some help here. However I read Hadley's piece on ``testthat`` and
think I could implement some simple tests like checking against creation of
empty data frames etc.

> * You have a circularity problem in your dependencies. DoOR.data is in DEPENDS
in this pkg, and DoOR.functions is in Suggests for the DoOR.data package - why
is DoOR.data in Depends? there's very few cases in which that's needed, is it
here?

True, in an earlier mail you suggested to put the data package into DEPENDS, I
had it in SUGGESTS before. The reason being that ``DoOR.functions`` depends on
some of the data provided in the ``DoOR.data`` package. This is also the reason
for the undefined global functions or variables below.

> * R check reveals many problems: you need to Rbuildignore
.travis.yml/CHANGELOG.md (but you should also have a NEWS.md or NEWS as that's
more typical of R packages),

ok

> * don't assign variables to the global environment
> * see ‘Good practice’ in ‘?data’ for help on data objects

As discussed in https://github.com/ropensci/onboarding/issues/35#issuecomment-218391691 this is how the package was initially written: "Load data into the
global environment so that it is easily visible and edible for the user". I am
looking into ways to rewrite that (started a branch for that). But this
is not simple as I might need to write lots of helper functions for the user to easily interact
with the data.

As described, my workaround for now was to create a user interaction upon
package loading. I prompt a warning message and the user has to explicitly type
"yes" to load the data.

> * you have many Undefined global functions or variables

These are from the ``DoOR.data`` package and that is why I put it to DEPENDS.
Would putting it to SUGGESTS and the using ``DoOR.data::`` work? But then if
e.g. a user modified version of ``door_response_matrix`` exists in the global
environment, the function should use this version and not pick the one from the
``DoOR.data`` package, which it would using the ``package::`` annotation, right?

> * A LOT OF warnings about "no visible binding for global variable' - fix those

Many of the arise from using ``ggplot`` functions. I just learned about:

```r
## quiets concerns of R CMD check re: the .'s that appear in pipelines
if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
```
here https://github.com/STAT545-UBC/Discussion/issues/451

#### summary
Points 1,3 & 7 I think I can fix. I would need some advice to fix
the circularity in 2/6. For point 4/5 I do not see an easy fix for now and I
hoped the warning and explicit user interaction will be a sufficient
workaround for now.
