### 2017-04-19 #34 @sckott
Let us know if you have testing questions, we do want them before accepting this pkg

I'll take a look at your dependency issue and global variables

> I just learned about: ...  utils::globalVariables

right, that should work
