### 2017-04-20 cc02e7c
added unit testing
