+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(DoOR.data)
+
+test_check("DoOR.data")
