+++ man/Or30a.Rd
@@ -21,4 +21,3 @@ information on the respective sources of the data in door_dataset_info.
 \keyword{Or30a}
 \keyword{dataset}
 \keyword{responding_unit}
-
