+++ man/Or2a.Rd
@@ -22,4 +22,3 @@ information on the respective sources of the data in door_dataset_info.
 \keyword{Or2a}
 \keyword{dataset}
 \keyword{responding_unit}
-
