+++ man/Or85b.Rd
@@ -26,4 +26,3 @@ information on the respective sources of the data in door_dataset_info.
 \keyword{Or85b}
 \keyword{dataset}
 \keyword{responding_unit}
-
