+++ man/Or33a.Rd
@@ -19,4 +19,3 @@ information on the respective sources of the data in door_dataset_info.
 \keyword{Or33a}
 \keyword{dataset}
 \keyword{responding_unit}
-
