+++ man/load_door_data.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/load_door_data.R
 \name{load_door_data}
-\alias{loadData}
 \alias{load_door_data}
+\alias{loadData}
 \title{load_door_data}
 \usage{
 load_door_data(nointeraction = FALSE)
@@ -29,4 +29,3 @@ data(Or22a)
 \author{
 Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
 }
-
