+++ man/Or92a.Rd
@@ -24,4 +24,3 @@ information on the respective sources of the data in door_dataset_info.
 \keyword{Or92a}
 \keyword{dataset}
 \keyword{responding_unit}
-
