+++ man/Or71a.Rd
@@ -22,4 +22,3 @@ information on the respective sources of the data in door_dataset_info.
 \keyword{Or71a}
 \keyword{dataset}
 \keyword{responding_unit}
-
