+++ man/Or85f.Rd
@@ -20,4 +20,3 @@ information on the respective sources of the data in door_dataset_info.
 \keyword{Or85f}
 \keyword{dataset}
 \keyword{responding_unit}
-
