### 2017-04-25 #34 @sckott
was checking, but it looks like some changes i asked for above are not done yet (e.g., `.travis.yml` should be put in `.Rbuildignore`) - i'll hold off until then
