### 2017-07-15 #34 @Dahaniel
@sckott I implemented unit testing in the data package, for the functions package I scheduled implementation for 1st half of August.
