### 2017-08-11 #34 @sckott
just checking in @Dahaniel - i'm waiting on this correct:

>  for the functions package I scheduled implementation for 1st half of August.
