### 2017-08-25 c07935e
moved CHANGELOG to NEWS.md
