+++ CHANGELOG.md
@@ -1,35 +1,0 @@
-# Change Log
-
-## [v2.0.1](https://github.com/Dahaniel/DoOR.data/tree/v2.0.1) (2017-03-07)
-[Full Changelog](https://github.com/Dahaniel/DoOR.data/compare/v2.0.0...v2.0.1)
-
-**Closed issues:**
-
-- replace '.' with '\_' in data names [\#19](https://github.com/Dahaniel/DoOR.data/issues/19)
-- moved all urls to url field in DESCRIPTION [\#18](https://github.com/Dahaniel/DoOR.data/issues/18)
-- add warning message that data is loaded into current workspace and force user feedback [\#17](https://github.com/Dahaniel/DoOR.data/issues/17)
-- rename `loadData` to `load\_door\_data\(\)` to be more descriptive [\#16](https://github.com/Dahaniel/DoOR.data/issues/16)
-- write documentation for data sets [\#15](https://github.com/Dahaniel/DoOR.data/issues/15)
-- revert to store data as text rather than binary [\#14](https://github.com/Dahaniel/DoOR.data/issues/14)
-
-## [v2.0.0](https://github.com/Dahaniel/DoOR.data/tree/v2.0.0) (2016-02-25)
-[Full Changelog](https://github.com/Dahaniel/DoOR.data/compare/v1.0.2...v2.0.0)
-
-**Fixed bugs:**
-
-- rename ab3B \> Or85b in DoOR.mappings [\#7](https://github.com/Dahaniel/DoOR.data/issues/7)
-
-**Closed issues:**
-
-- setf SFR == 0 for all Ca²⁺ imaging studies [\#4](https://github.com/Dahaniel/DoOR.data/issues/4)
-- set SFR == 0 for all studies that subtracted SFR but did not report it [\#3](https://github.com/Dahaniel/DoOR.data/issues/3)
-- wrong value in Hallem.2006.EN [\#2](https://github.com/Dahaniel/DoOR.data/issues/2)
-- duplicated odor [\#1](https://github.com/Dahaniel/DoOR.data/issues/1)
-
-## [v1.0.2](https://github.com/Dahaniel/DoOR.data/tree/v1.0.2) (2015-04-17)
-[Full Changelog](https://github.com/Dahaniel/DoOR.data/compare/v1.0...v1.0.2)
-
-## [v1.0](https://github.com/Dahaniel/DoOR.data/tree/v1.0) (2014-12-04)
-
-
-\* *This Change Log was automatically generated by [github_changelog_generator](https://github.com/skywinder/Github-Changelog-Generator)*
