+++ DESCRIPTION
@@ -2,8 +2,8 @@ Package: DoOR.data
 Type: Package
 Title: Integrating Heterogeneous Odorant Response Data into a Common Response
     Model: A DoOR to the Complete Olfactome
-Version: 2.0.1
-Date: 2017-03-07
+Version: 2.0.1.9000
+Date: 2017-08-25
 Authors@R: c(
     person("Daniel", "Münch", email = "daniel.muench@uni-konstanz.de", role = c("aut", "cre")),
     person("C. Giovanni", "Galizia", email = "giovanni.galizia@uni-konstanz.de", role = "aut"), 
@@ -19,7 +19,7 @@ Encoding: UTF-8
 Imports:
     utils
 Depends:
-    R (>= 3.2.0)
+    R (>= 3.4.1)
 Suggests:
     DoOR.functions (>= 2.0.1),
     testthat
