### 2017-08-25 #34 @Dahaniel
Ok, should be done.

I fixed:
- [x] added basic unit tests to both packages
- [x] fixed the 'no visible binding for global variable'
- [x] moved CHANGELOG to NEWS.md
- [x] added entries to .Rbuildignore

there remains:
- [ ] a NOTE about loading data to the global environment  that will stay for now (see https://github.com/ropensci/onboarding/issues/35#issuecomment-218391691)
- [ ] circularity: Not sure what is the appropriate solution: I kept `DoOR.data` in Depends as the algorithms in `DoOR.functions` need data from `DoOR.data` to run (vignettes and examples would not work without). `DoOR.functions` is in suggests in `DoOR.data` as  the two packages basically belong together.
