### 2017-09-05 #34 @sckott
@Dahaniel

For the data issue: isn't it possible to do this approach:

- use the default data.frame in `DooR.data` package (NOT loaded into the global namespace, but only in the package namespace)
- optionally, the user can pass in a modified version of the data.frame to a parameter in the function; overriding the default data.frame
- for the user to get access to the data.frame to modify it, they could simply load via `data(..)` i think

this should get around  the problem of the package loading data into the global namespace

i think we do need to solve this issue as CRAN will likely not accept it as is

---

I think the circularity thing is actually okay - as long as one has the other in Suggests
