### 2017-09-06 #34 @Dahaniel
Thanks @sckott , how can I reproduce the good practice output you posted here?
