### 2017-09-06 #34 @sckott
install `devtools::install_github("MangoTheCat/goodpractice")` then run `.Sys.setenv(NOT_CRAN = "true"); x = goodpractice::gp(); x`
