### 2017-10-09 74bf4fc
sapply > vapply
