### 2017-10-14 #34 @Dahaniel
Hi @sckott , sorry, busy as usual.
I actually startet working on fixing the "good practice" warnings and am done with the data package.
Regarding the data issue: What makes it complicated is that it is actually not a single data.frame which is accessed but many and that there are many functions that need to be updated. So I think this will take some effort for writing as well as testing.
If ok for you I would rather fix the GP issues now for the onboarding and then have CRAN compatibility as a new milestone afterwards.
