### 2017-11-10 #34 @sckott
@Dahaniel Okay, did you get to fixing these goodpractice problems yet?
