### 2017-11-17 #34 @Dahaniel
@sckott Should be fixed for both packages, shortened code-lines, replaced sapply with vapply etc...
