### 2018-02-11 #34 @Dahaniel
That's awesome news! I'll see to your points in the coming days and I am happy to write a blog-post as well.
