### 2018-02-11 78da9f7
added ROpenSci banner
