### 2018-02-12 #34 @Dahaniel
@sckott where should I look for the invitation to the ROpenSci group? Could not find anything on github or in my mails.
