### 2018-02-12 #34 @stefaniebutland
Hello @Dahaniel. Great to hear you're interested in writing a blog post about this. Here are some technical and editorial notes for contributing a post: https://github.com/ropensci/roweb2#contributing-a-blog-post

We ask that you submit a draft via pull request at least one week prior to the planned publication date. I have two dates open right now. Take your pick:
1) Tues Feb 20 (draft due Wed Feb 14 if possible)
2) Tues Mar 20 (draft due Tues Mar 13)
