+++ README.md
@@ -1,10 +1,10 @@
 DoOR.data
 =========
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.375618.svg)](http://dx.doi.org/10.5281/zenodo.375618)
-[![Travis-CI Build Status](https://travis-ci.org/Dahaniel/DoOR.data.svg?branch=master)](https://travis-ci.org/Dahaniel/DoOR.data)
+[![Travis-CI Build Status](https://travis-ci.org/ropensci/DoOR.data.svg?branch=master)](https://travis-ci.org/ropensci/DoOR.data)
 
 
-R package containing the data for the Database of Odor Responses. The corresponding functions package can be found at [https://github.com/Dahaniel/DoOR.functions](https://github.com/Dahaniel/DoOR.functions).
+R package containing the data for the Database of Odor Responses. The corresponding functions package can be found at [https://github.com/ropensci/DoOR.functions](https://github.com/Dahaniel/DoOR.functions).
 
 ## The DoOR Project
 Find more information, precompiled R-packages and a interactive web-version of the DoOR-Database at: **[http://neuro.uni.kn/DoOR](http://neuro.uni.kn/DoOR)**
@@ -18,10 +18,10 @@ install.packages("devtools")
 library(devtools)
 
 # install DoOR.data 2.0.1
-install_github("dahaniel/DoOR.data", ref="v2.0.1")
+install_github("ropensci/DoOR.data", ref="v2.0.1")
 
 # or install the latest version available on Github
-install_github("dahaniel/DoOR.data")
+install_github("ropensci/DoOR.data")
 ```
 
 ## Publications
