### 2018-02-24 #34 @Dahaniel
@stefaniebutland is there a guideline or a specific structure that I should follow when writing the blog post?
