### 2018-02-25 #34 @stefaniebutland
If it would help, I could give you feedback on an outline where you identify audience, take home message and key points you want to get across. No obligation to do this
